Fix TUI regressions found in adversarial review
The onboarding refusal was built on a false premise — the dashboard's own
chat route documents that onboarding is deliberately NOT born at project
bootstrap and births it on open. Restore create-if-unborn with the real birth
batch (the approve-core cross-import pattern). Also: TUI-side linear-backoff
retry for transient non-transport failures (a terminal has no refresh
button), per-cycle/per-send agent stub disposal on the process-long socket,
react + react-query inlined into the CLI/TUI bundle (peers only for library
entries), probe-credentials regression test, parked-aware configure guard
message, react-doctor fixes (children as createElement args, chat-view/
chat-colors split).
