+++ packages/iterate/src/stream-tui/chat-colors.ts
@@ -1,0 +1,14 @@
+/** The agent chat TUI palette, shared by the app shell and the view components. */
+export const COLORS = {
+  bg: "#0b0f14",
+  surface: "#27272a",
+  border: "#3f3f46",
+  accent: "#22c55e",
+  warning: "#facc15",
+  danger: "#ef4444",
+  text: "#e5e7eb",
+  textSecondary: "#9ca3af",
+  textBody: "#d1d5db",
+  textMuted: "#6b7280",
+  agent: "#a78bfa",
+} as const;
